Refactor Armor and Weapon SO. Started creating CAD Grid Menu
